refactor(console): rename and move console class
