+++ src/main/java/racingcar/Game.java
@@ -1,24 +1,26 @@
 package racingcar;
 
-import utils.RandomUtils;
+import racingcar.consol.InputConsole;
+import racingcar.consol.OutputConsole;
 
-import java.util.List;
 import java.util.Scanner;
 
 public class Game {
     private static final int START_INCLUSIVE = 0;
     private static final int END_INCLUSIVE = 9;
     private Car[] cars;
-    private Player player;
-    private Printer printer;
+    private InputConsole inputConsole;
+    private OutputConsole outputConsole;
 
     public Game(Scanner scanner) {
-        player = new Player(scanner);
-        printer = new Printer();
+        inputConsole = new InputConsole(scanner);
+        outputConsole = new OutputConsole();
     }
 
     /* 경주할 자동차 이름, 시도할 횟수 입력 받기 */
     public void start() {
+        outputConsole.printInputCarNameMessage();
+        inputConsole.getCarNameArray();
     }
 
     /* 각 차수별로 각 자동차를 전진 또는 멈춤하고 실행 결과 출력 */
@@ -27,9 +29,6 @@ public class Game {
 
     /* 최종 우승자 뽑고 출력 */
     public void finish() {
-    }
-
-    public void splitNamesWithComma() {
     }
 
     public void createRandomNumber() {
