+++ src/main/java/racingcar/consol/OutputConsole.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.consol;
 
-public class Printer {
+public class OutputConsole {
     private static final String INPUT_CAR_NAME_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
     private static final String INPUT_TRY_NUMBER_MESSAGE = "시도할 횟수는 몇회인가요?";
     private static final String MOVING_RESULT_HEADER_MESSAGE = "실행 결과";
@@ -8,12 +8,15 @@ public class Printer {
     private static final String OUTPUT_FINAL_WINNER_MESSAGE = "최종 우승자: ";
 
     public void printInputCarNameMessage() {
+        System.out.println(INPUT_CAR_NAME_MESSAGE);
     }
 
     public void printInputTryNumberMessage() {
+        System.out.println(INPUT_TRY_NUMBER_MESSAGE);
     }
 
-    public void printResultPositionHeader() {
+    public void printMovingResultHeader() {
+        System.out.println(MOVING_RESULT_HEADER_MESSAGE);
     }
 
     public void printOutputEachResultPosition() {
